Legacy_AEFI_Acquisition/getE3D/requirements.txt
pyserial>=3.5
PyQt5>=5.15.4
pyqtgraph>=0.12.0
numpy>=1.20.0 